module: module_adapter: Move module_interface header to module
Moved header files to the module directory to separate an shared interface
used by sof and native loadable modules.
